fix: dynamic page build
[#143] correção rota "/construtoras/cadastrar"

https://jira.arinterface.com.br/work_packages/143
